New version v1.4: Adopted new course and assigned new teacher
This commit updates the CSMI program by replacing two existing courses - Networking and Pre and Post-Processing in Scientific Computing - with a new course titled Network-Integrated Pre and Post-Processing in Scientific Computing. The teacher for Algorithmics and Graphs has been updated to Y. Bugeaud, and colleagues external to the UFR are now indicated by prefixing their names with *Ext.* in the course description. More details on the changes can be found in the ChangeLog.